+++ src/menu.h
@@ -11,8 +11,10 @@
 
 #include "button.h"
 #include "fonts.h"
+#include "history.h"
 #include "icons.h"
 #include "model.h"
+#include "util.h"
 
 #define T(x) GetLangText(x)
 
@@ -23,12 +25,14 @@ namespace taranis {
 
 enum menu_item_index {
   MENU_ITEM_REFRESH = 100,
-  MENU_ITEM_UNIT_SYSTEM = 200,
-  MENU_ITEM_ABOUT = 201,
+  MENU_ITEM_LOCATION_HISTORY = 200,
+  MENU_ITEM_UNIT_SYSTEM = 201,
+  MENU_ITEM_ABOUT = 202,
   MENU_ITEM_QUIT = 300,
   MENU_ITEM_UNIT_SYSTEM_STANDARD = 600,
   MENU_ITEM_UNIT_SYSTEM_METRIC = 601,
   MENU_ITEM_UNIT_SYSTEM_IMPERIAL = 602,
+  MENU_ITEM_EMPTY_LOCATION_HISTORY = 800,
 };
 
 class MenuButton : public Button {
@@ -45,15 +49,20 @@ public:
                           imenu{0, 0, nullptr, nullptr}},
         items{imenu{ITEM_ACTIVE, taranis::MENU_ITEM_REFRESH,
                     const_cast<char *>(T("Refresh")), nullptr},
+              imenu{ITEM_SUBMENU, taranis::MENU_ITEM_LOCATION_HISTORY,
+                    const_cast<char *>(T("Locations")),
+                    this->location_history_items.data()},
               imenu{ITEM_SUBMENU, taranis::MENU_ITEM_UNIT_SYSTEM,
                     const_cast<char *>(T("Units")),
-                    const_cast<imenu *>(unit_system_items.data())},
+                    const_cast<imenu *>(this->unit_system_items.data())},
               imenu{ITEM_ACTIVE, taranis::MENU_ITEM_ABOUT,
                     const_cast<char *>(T("About…")), nullptr},
               imenu{ITEM_ACTIVE, taranis::MENU_ITEM_QUIT,
                     const_cast<char *>(T("Quit")), nullptr},
               imenu{0, 0, nullptr, nullptr}},
-        model{model}, font{fonts->get_normal_font()} {}
+        model{model}, font{fonts->get_normal_font()} {
+    this->update_location_history_items();
+  }
 
   void set_menu_handler(iv_menuhandler handler) {
     this->menu_handler = handler;
@@ -75,7 +84,10 @@ private:
   static const int icon_size{70};
 
   std::array<imenu, 4> unit_system_items;
-  const std::array<imenu, 5> items;
+  std::array<imenu, LocationHistory::max_size + 1> location_history_items;
+  const std::array<imenu, 6> items;
+
+  std::vector<std::string> location_full_names;
 
   std::shared_ptr<Model> model;
 
@@ -99,19 +111,56 @@ private:
     return {pos_x, pos_y};
   }
 
-  void update_unit_system_bullet() const {
+  void update_location_history_items() {
+    this->location_full_names.clear();
+
+    size_t index = 0;
+    auto &location_history = this->model->location_history;
+    if (location_history.empty()) {
+      this->location_history_items.at(index) =
+          imenu{ITEM_INACTIVE, taranis::MENU_ITEM_EMPTY_LOCATION_HISTORY,
+                const_cast<char *>(T("Empty")), nullptr};
+      ++index;
+    } else {
+      for (const auto &location : location_history) {
+        this->location_full_names.push_back(format_location(location));
+
+        const short item_index =
+            taranis::MENU_ITEM_EMPTY_LOCATION_HISTORY + index + 1;
+        this->location_history_items.at(index) =
+            imenu{ITEM_ACTIVE, item_index, nullptr, nullptr};
+
+        ++index;
+        if (index >= this->location_history_items.size() - 1) {
+          break;
+        }
+      }
+    }
+    while (index < this->location_history_items.size()) {
+      this->location_history_items.at(index) = imenu{0, 0, nullptr, nullptr};
+      ++index;
+    }
+    index = 0;
+    for (const auto &item_name : this->location_full_names) {
+      this->location_history_items.at(index).text =
+          const_cast<char *>(item_name.c_str());
+      ++index;
+    }
+  }
+
+  void update_unit_system_bullet() {
     if (this->model->unit_system == UnitSystem::standard) {
-      const_cast<imenu *>(&this->unit_system_items[0])->type = ITEM_BULLET;
-      const_cast<imenu *>(&this->unit_system_items[1])->type = ITEM_ACTIVE;
-      const_cast<imenu *>(&this->unit_system_items[2])->type = ITEM_ACTIVE;
+      this->unit_system_items[0].type = ITEM_BULLET;
+      this->unit_system_items[1].type = ITEM_ACTIVE;
+      this->unit_system_items[2].type = ITEM_ACTIVE;
     } else if (this->model->unit_system == UnitSystem::metric) {
-      const_cast<imenu *>(&this->unit_system_items[0])->type = ITEM_ACTIVE;
-      const_cast<imenu *>(&this->unit_system_items[1])->type = ITEM_BULLET;
-      const_cast<imenu *>(&this->unit_system_items[2])->type = ITEM_ACTIVE;
+      this->unit_system_items[0].type = ITEM_ACTIVE;
+      this->unit_system_items[1].type = ITEM_BULLET;
+      this->unit_system_items[2].type = ITEM_ACTIVE;
     } else if (this->model->unit_system == UnitSystem::imperial) {
-      const_cast<imenu *>(&this->unit_system_items[0])->type = ITEM_ACTIVE;
-      const_cast<imenu *>(&this->unit_system_items[1])->type = ITEM_ACTIVE;
-      const_cast<imenu *>(&this->unit_system_items[2])->type = ITEM_BULLET;
+      this->unit_system_items[0].type = ITEM_ACTIVE;
+      this->unit_system_items[1].type = ITEM_ACTIVE;
+      this->unit_system_items[2].type = ITEM_BULLET;
     }
   }
 
@@ -119,6 +168,7 @@ private:
     if (not this->menu_handler)
       return;
 
+    this->update_location_history_items();
     this->update_unit_system_bullet();
 
     const auto &[pos_x, pos_y] = this->get_menu_position();